refactor of testkit
