+++ crates/aos-examples/Cargo.toml
@@ -12,7 +12,7 @@ aos-store = { path = "../aos-store" }
 aos-air-exec = { path = "../aos-air-exec" }
 aos-wasm-abi = { path = "../aos-wasm-abi" }
 aos-effects = { path = "../aos-effects" }
-aos-testkit = { path = "../aos-testkit" }
+aos-host = { path = "../aos-host", features = ["test-fixtures"] }
 aos-wasm-build = { path = "../aos-wasm-build" }
 clap = { version = "4", features = ["derive"] }
 indexmap = "2"
