+++ crates/aos-examples/src/examples/fetch_notify.rs
@@ -3,7 +3,7 @@ use std::path::Path;
 use anyhow::{Result, anyhow};
 use serde::{Deserialize, Serialize};
 
-use crate::support::http_harness::{HttpHarness, MockHttpResponse};
+use aos_host::adapters::mock::{MockHttpHarness, MockHttpResponse};
 use crate::support::reducer_harness::{ExampleReducerHarness, HarnessConfig};
 
 const REDUCER_NAME: &str = "demo/FetchNotify@1";
@@ -49,7 +49,7 @@ pub fn run(example_root: &Path) -> Result<()> {
     println!("     start fetch → url={url} method={method}");
     run.submit_event(&start_event)?;
 
-    let mut http = HttpHarness::new();
+    let mut http = MockHttpHarness::new();
     let requests = http.collect_requests(run.kernel_mut())?;
     if requests.len() != 1 {
         return Err(anyhow!(
