+++ crates/aos-examples/src/support/mod.rs
@@ -1,4 +1,3 @@
-pub mod http_harness;
 pub mod manifest_loader;
 pub mod reducer_harness;
 pub(crate) mod util;
