+++ crates/aos-host/src/adapters/mock.rs
@@ -1,19 +1,204 @@
 //! Mock adapters for testing.
 //!
-//! This module provides mock implementations of effect adapters for use in tests.
+//! This module provides mock harnesses for intercepting and responding to effects in tests:
+//!
+//! - [`MockHttpHarness`]: Intercepts `http.request` effects and provides mock responses
+//! - [`MockLlmHarness`]: Intercepts `llm.generate` effects and provides synthetic responses
 
 use std::sync::Arc;
 
 use anyhow::{Context, Result, anyhow};
-use aos_air_exec::Value as ExprValue;
+use aos_air_exec::{Value as ExprValue, ValueKey};
 use aos_air_types::HashRef;
 use aos_cbor::Hash;
-use aos_effects::builtins::LlmGenerateParams;
+use aos_effects::builtins::{HeaderMap, HttpRequestParams, LlmGenerateParams};
 use aos_effects::{EffectIntent, EffectKind, EffectReceipt, ReceiptStatus};
 use aos_kernel::Kernel;
 use aos_store::Store;
 use sha2::{Digest, Sha256};
 use tracing::debug;
+
+// ---------------------------------------------------------------------------
+// MockHttpHarness: HTTP effect interception
+// ---------------------------------------------------------------------------
+
+const MOCK_HTTP_ADAPTER_ID: &str = "http.mock";
+
+/// Context for an HTTP request intercepted by the mock harness.
+#[derive(Debug, Clone)]
+pub struct HttpRequestContext {
+    pub intent: EffectIntent,
+    pub params: HttpRequestParams,
+}
+
+/// Mock HTTP response for testing.
+#[derive(Debug, Clone)]
+pub struct MockHttpResponse {
+    pub status: i64,
+    pub headers: HeaderMap,
+    pub body: String,
+}
+
+impl MockHttpResponse {
+    /// Create a JSON response with the given status and body.
+    pub fn json(status: i64, body: impl Into<String>) -> Self {
+        let mut headers = HeaderMap::new();
+        headers.insert(
+            "content-type".into(),
+            "application/json; charset=utf-8".into(),
+        );
+        Self {
+            status,
+            headers,
+            body: body.into(),
+        }
+    }
+
+    /// Create a plain text response.
+    pub fn text(status: i64, body: impl Into<String>) -> Self {
+        let mut headers = HeaderMap::new();
+        headers.insert("content-type".into(), "text/plain; charset=utf-8".into());
+        Self {
+            status,
+            headers,
+            body: body.into(),
+        }
+    }
+}
+
+/// Mock HTTP harness for testing HTTP effect flows.
+///
+/// This harness intercepts `http.request` effects and allows tests to
+/// provide mock responses.
+pub struct MockHttpHarness;
+
+impl MockHttpHarness {
+    pub fn new() -> Self {
+        Self
+    }
+
+    /// Collect all pending HTTP requests from the kernel.
+    pub fn collect_requests<S: Store + 'static>(
+        &mut self,
+        kernel: &mut Kernel<S>,
+    ) -> Result<Vec<HttpRequestContext>> {
+        let mut out = Vec::new();
+        loop {
+            let intents = kernel.drain_effects();
+            if intents.is_empty() {
+                break;
+            }
+            for intent in intents {
+                match intent.kind.as_str() {
+                    EffectKind::HTTP_REQUEST => {
+                        let params: HttpRequestParams =
+                            serde_cbor::from_slice(&intent.params_cbor)
+                                .context("decode http request params")?;
+                        out.push(HttpRequestContext { intent, params });
+                    }
+                    other => {
+                        return Err(anyhow!("unexpected effect kind {other}"));
+                    }
+                }
+            }
+        }
+        Ok(out)
+    }
+
+    /// Respond to an HTTP request with a mock response.
+    pub fn respond_with<S: Store + 'static>(
+        &self,
+        kernel: &mut Kernel<S>,
+        ctx: HttpRequestContext,
+        response: MockHttpResponse,
+    ) -> Result<()> {
+        self.respond_with_body(kernel, None::<&S>, ctx, response)
+    }
+
+    /// Respond to an HTTP request, optionally storing the body in a store.
+    pub fn respond_with_body<S: Store + 'static>(
+        &self,
+        kernel: &mut Kernel<S>,
+        store: Option<&impl Store>,
+        ctx: HttpRequestContext,
+        response: MockHttpResponse,
+    ) -> Result<()> {
+        let receipt_value =
+            build_http_receipt_value(response.status, &response.headers, response.body, store)?;
+        let receipt = EffectReceipt {
+            intent_hash: ctx.intent.intent_hash,
+            adapter_id: MOCK_HTTP_ADAPTER_ID.into(),
+            status: ReceiptStatus::Ok,
+            payload_cbor: serde_cbor::to_vec(&receipt_value)?,
+            cost_cents: Some(0),
+            signature: vec![0; 64],
+        };
+        kernel.handle_receipt(receipt)?;
+        kernel.tick_until_idle()?;
+        Ok(())
+    }
+}
+
+impl Default for MockHttpHarness {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+fn build_http_receipt_value(
+    status: i64,
+    headers: &HeaderMap,
+    body: String,
+    store: Option<&impl Store>,
+) -> Result<ExprValue> {
+    let mut record = indexmap::IndexMap::new();
+    record.insert("status".into(), ExprValue::Int(status));
+    record.insert("headers".into(), headers_to_value(&redact_headers(headers)));
+    record.insert("body_preview".into(), ExprValue::Text(body.clone()));
+    if let Some(store) = store {
+        let hash = store
+            .put_blob(body.as_bytes())
+            .context("store http response body")?;
+        record.insert("body_ref".into(), ExprValue::Text(hash.to_hex()));
+    }
+    let mut timings = indexmap::IndexMap::new();
+    timings.insert("start_ns".into(), ExprValue::Nat(10));
+    timings.insert("end_ns".into(), ExprValue::Nat(20));
+    record.insert("timings".into(), ExprValue::Record(timings));
+    record.insert(
+        "adapter_id".into(),
+        ExprValue::Text(MOCK_HTTP_ADAPTER_ID.into()),
+    );
+    Ok(ExprValue::Record(record))
+}
+
+fn headers_to_value(headers: &HeaderMap) -> ExprValue {
+    let mut map = aos_air_exec::ValueMap::new();
+    for (key, value) in headers {
+        map.insert(ValueKey::Text(key.clone()), ExprValue::Text(value.clone()));
+    }
+    ExprValue::Map(map)
+}
+
+fn redact_headers(headers: &HeaderMap) -> HeaderMap {
+    let mut redacted = HeaderMap::new();
+    for (k, v) in headers {
+        let lower = k.to_ascii_lowercase();
+        if matches!(
+            lower.as_str(),
+            "authorization" | "proxy-authorization" | "x-api-key" | "api-key"
+        ) {
+            redacted.insert(k.clone(), "<redacted>".into());
+        } else {
+            redacted.insert(k.clone(), v.clone());
+        }
+    }
+    redacted
+}
+
+// ---------------------------------------------------------------------------
+// MockLlmHarness: LLM effect interception
+// ---------------------------------------------------------------------------
 
 const MOCK_LLM_ADAPTER_ID: &str = "llm.mock";
