--- crates/aos-testkit/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "aos-testkit"
-version = "0.1.0"
-edition = "2024"
-publish = false
-description = "Re-export shim for aos-host test fixtures (deprecated - use aos-host directly)"
-
-[dependencies]
-# All functionality is now in aos-host with test-fixtures feature
-aos-host = { path = "../aos-host", features = ["test-fixtures"] }
